moved and renamed post interface
