--- frontend/dogplus-frontend/src/assets/interfaces.ts
@@ -1,11 +1,0 @@
-
-export interface Post {
-    id: number;
-    author: string;
-    profile_pic: string;
-    like_count: number;
-    comment_count: number;
-    date: string;
-    text: string;
-    image?: string;
-  }
